feat(oracle_rewrite): Tests structure
