--- crates/accounts/tests/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(test)]
-pub mod test_oracle;
-
-#[cfg(test)]
-pub mod test_publisher;
